[FIX] 콘텐츠, 카테고리 중복 삭제 방지 로직 추가 (#319)
* Fix: 콘텐츠 존재 여부 삭제 이전에 확인

* Fix: 카테고리 삭제 전 삭제 가능 여부 확인

* Chore: 누락된 인덴트 수정

+++ functions/api/routes/content/contentDELETE.js
@@ -35,15 +35,14 @@ module.exports = asyncWrapper(async (req, res) => {
   dayjs.tz.setDefault('Asia/Seoul');
 
   const content = await contentDB.getContentById(dbConnection, contentId);
+  if (!content) {
+    return res.status(statusCode.NOT_FOUND).send(util.fail(statusCode.NOT_FOUND, responseMessage.NO_CONTENT));
+  }
   if (content.userId !== userId) {
     return res.status(statusCode.FORBIDDEN).send(util.fail(statusCode.FORBIDDEN, responseMessage.FORBIDDEN));
   }
 
-  const deletedContent = await contentDB.deleteContent(dbConnection, contentId, userId);
-  if (!deletedContent) {
-    // 대상 콘텐츠가 없는 경우, 콘텐츠 삭제 실패
-    return res.status(statusCode.NOT_FOUND).send(util.fail(statusCode.NOT_FOUND, responseMessage.NO_CONTENT));
-  }
+  await contentDB.deleteContent(dbConnection, contentId, userId);
 
   if (content.isNotified && content.notificationTime > dayjs().tz().$d) {
     // 알림 예정 일 때 푸시서버에서도 삭제
